+++ changelog.md
@@ -4,6 +4,7 @@
 
 - Solved [issue 5](https://github.com/p4535992/foundryvtt-arms-reach/issues/5)
 - Add more Internationalization
+- Add Lock Picking Sound
 
 # 1.0.8 (Breaking Version)